Split messaging test setup to its own module
